Added makefile to run the compilation to executable with the standalone c function of the assembler, for both release and debug. Note that this requires the lt64-asm-0.0.3-1.jar currently. Also, added a .gitignore.
Created a starting place for the program and a simple module with a
temporary subroutine. It should be possible towork form here to start
creating the actual interpreter. The starting point should be the reader
and printer, which I beleive will require the allocator to create in
memory representations of the read in data. Once it is possible to read
and print simple s-expressions then the eval function can be made.