Started working on the allocation subroutines. Then decided I needed some testing before they could be commited and ok. So wrote a little test runner and setup a single test for the first allocate routine.
It is not super easy, but basically a module of tests for a module can be
made and then those tests can be called in the runner. The test should
leave the expected and actual value to check on the stack. Then call the
is function to check if they are equal. If they are it will print PASSED
and if they are not it will print FAILED and the expected vs actual.
Finally, when the tests are over if there were any failures it will
print the number of them.

Obviously a test setup in an assembly language will not be full
featured. It could probably do a little more. Mainly it will be hard
with a large number of tests to find the failures and know what they
mean. So the convention will be to explain the test in a comment in the
test itself. Then if all tests for a module are wrapped nicely and
module tests suites are separated with titles it is possible to figure
out which test failed and check the test function. To make this work it
will also be best if modules and their test modules are kept smaller. If
they are too large then it will be really hard to track down failures.

The tests are really important because we all make mistakes with high
level langauges that are easy to use. This one is not too bad for a low
level language, but it is definatley easy to make mistakes.