Fixed an include in the test runner.
Setup a function for reading symbols when we know that the next is a
symbol. Right now it basically is a readline function. It is not tested
to ensure that the storage is correct, but it does read chars into the
buffer packed properly.

The next thing is to make it stop on any delimeter character.