Adjusted subroutine documentation to put the general description above the subroutine instead of inside it. I think this distracts less from the operations, everything is already a little cluttered.
Moved the setup for data structures out of the allocate page. The idea
is that each structure should have its own macros and procedures for
accessing its content that are separate from its implementation. This is
helpful since the internals are pretty low level and I don't want to be
pushing raw numbers with additions to addressed to load and store
values. It also makes testing them much easier.

Allocate now just gives the address of memory and moves the next open
pointer. The tests have been updated, and do rely on the raw size of the
objects, but that is ok for now.

Added a symbol object that knows how to setup a symbol with the correct
data and has macros to access and set the different members. The print
name section does not have a get and set as it will be set with memcopy
or strcopy. Maybe a setter could be made eventually that takes 2 memory
addresses, but for now the pointer can be accessed and used to call
memcopy to copy chars from the buffer of memory (:buf-to-mem).

Also added a couple of tests for the symbol and am setup to add the
rest.