Updated the read/form to return nil on error instead of halting so that the error does not crash the repl, just moves to next level. Note that this is not the correct behaviour if a file is being read for eval without the repl.
Updated the symbol char function to check for invalid symbol characters.
This is used in read/form to see if char can start a symbol. It does not
have to check for digits and parens because they will be checked before
this call in read/form. It also does not check for chars that will end a
symbol because they are checked to stop the symbol in other places.

Again all is tested informally because I do not have a way to test io
right now. Though some of them don't use io. So maybe there are a couple
missing tests.