Got a mostly working version of read/pair. Currently it does not work with nested pairs. Somehow it creates a circular list and print recurses to stack overflow.
Another strange problem has come up where the input eventually just
perma blocks. I have tried all kinds of things including adding
statements to the vm around reads. However, nothing seems to happen
after a certain point the program seems to want input, but no input will
cause it to move forward. It seems kind of like when input is waiting
for an int and nothing you enter makes a difference if it is not an int,
but in this case there is no character I can enter that will stop it,
and no output to indicate what part of the program might be doing it.
Perhaps there is some internal buffer that fills up, though it happens
too quickly for that.