Heavily rewrote the symbol set and print to use loops now that I have the rcount macros and loops are easier. This allows the set-name subroutine to zero additional bytes in the word. Might not be essential for now, but if the symbol is using reclaimed memory it will need to zero those extra words.
Also, added a symbol function to check if two names are equal. This
fixes the problem that was present for names not properly being seen as
equal because the buffer could have extra garbage when names are small
and other things have been read in the buffer. The method of just
comparing all 4 words would not work for this, and the use of streq
worked only if the names were 7 or less chars. Since I am already using
lots of space for the symbols and other memory allocations, it makes
sense to not force one of the bytes to be 0 to stop the equality checks.
It is probably a little less efficient than if I could hand it over to
the VM to check, but it is only 4 bytes being checked so the cost of
checking with a subroutine should be minimal (again I don't want to
ignore efficiency, but it is not the main goal of the project).