Added all the functionality for defines. As long as the interpreter is open defined values can be used. Tested defining and evaluating symbols to get their values and it works as expected.
Also added an error value
instead of just a printable name. This is part of the symlist and the
builtin env in the same way as T F NIL. However, I have not yet changed
the functions leaving :!error valued, and those handling them to use the
new symbol. It makes sense that if I want to return it from evaluations
etc that it should be an actual value and should be useable in the lisp
code, say in a condition to check for errors. Perhaps later it can even
be used with a list of data to return error details and pass them
around.