Added something to make it possible to have recursive lambdas defined. This is somewhat succesful in that everything else works. However, some issues arose. The main being that no complex recursive function does anything. A simple one can infinite loop, but with certain forms in it it will stop early and not work as expected. This could be a procedure application issue or another issue.
Trying to do the recursion without arithmetic operators meant doing a
simple list building function. The issue was that for some reason
printing a form like this (eq? NIL a) will recurse forever. I also do
not know what is behind this. It is not the issue that affects recursion
as I duplicated it in another test with no lambda or recursion. So it
needs to be investigated further and fixed.

There are some tests to demonstrate some of the behavior listed above,
but they are commented out for no to ensure everything else is working.