Fixed the printing issue by removing the old :!nil printing case and letting it just print the symbol. The same needs to be done for error, but it needs to wait until error is fixed and handled everywhere. Also, fixed an issue where printing dotted pairs with NIL at the end of the print pair were printing the NIL. This split is kind of awkward but it works for now.
Recursive functions now work as expected with define. The issue was the
argument evaluation subroutine. It was mutating the procedure body when
the function was recursive. This was removed, and was easier the new way
anyway, and now the recursion finishes properly.

The changes made here may work for let as well, but with a repl and
defines it is not so bad that let does not work for recursion. A new let
could be added, letrec, or just leave it with defines for now. I don't
think old lisp had recursion in lets, but I am not really sure I checked
that.

What needs to be done now is arithmetic and relational operators. Then
other than things like i/o it will be able to do many things. After that
the question is how to add more builtin functions without evaluating
them at startup.