Added a way to create subroutines that are not builtin the same way as the basic and special forms. This requires loading them, which is not what I wanted, but I cannot know the address of a subroutine ahead of time, unless the assembler is changed to allow labels in the static section. It works in the scope of this project and still probably wont create long startup times, as the size of everything is still so small.
Added the add function. Cannot use the symbol right now because it is
expected to start a number, same with negative sign. So I will probably
use word names for the general arithmetic operations.

The main thing that seems gross about the addition function is that
because it has to deal with both int and fixed it has to be a little
complicated. Once a fixed is seen then they all have to be converted to
fixed if needed. Then the loop has to be run expecting that at the end
we will be creating a fixed point number. This could probably be done
with a flag somewhere on the stack instead, but it is easier to split it
up. The thing about that that I do not like the most is that each
function will differ only in what operators are used in each function. I
think it could be possible to set the op code as a value and push it so
the same function could be used a bunch, but it seems more trouble than
it is worth in some ways. However, without some scheme like this the
core lib is going to get really big for such simple functions.

Though perhaps only add sub mult and div will work like this. The rest
can be binary operators and just take two values. REMINDER that there is
nothing that forces this. If the function is given too many the extra
are ignored, but too few will likely cause a crash.