Added an updated version of the vm that saves the first cli argument to the buffer. This is accessed to open a file and read the first form and close the file. Then the form can be evaluated and the result displayed. The next thing is that it is necessary to add some basic i/o to allow reading and printing during program execution. Once this is done the print/form can be removed.
After this is accomplished the only thing stopping using the programs
for something like contest programming is somehow getting the program
data into the cfile. This can probably be done in a roundabout way by
writing a function that prints the contents of memory so that it can be
included in the standalone vm file. The easiest way would be to print
all memory and just concat it onto the end with a shell script. The vm
does not have the capability for me to do this from in a ltsp program.
So the goal will be to dump memory into a file and then use that to
create the cfile.

Also, without proper garbage collection and allocation the size of
programs that will be able to run will be limited.