Added a load function that takes a name from the buffer and loads all the forms in it. This is not yet useful for the repl as it is not really possible to get the filename in the buffer. Later load should take a string argument that is the filename, that can be internally loaded into the buffer for use with open.
Now it is useful to write a program that has multiple forms. This has
replaced the previous operations that just run a form in a file. Now all
forms in a file are run, without any printing. Then once done we check
for a define *main* and if it is there we run it as a function that
takes no arguments. This means that if there is not defined main then
an error will be printed and nothing will happen. If the main is not a
lambda it will fail when we try to apply it to the arguments.

These are not fully error checked etc. so they may not work in a perfect
way yet in all situations. Also, currently the main result is printed to
show that it was executed properly, but once print functions are
available then this will not happen, and only program defined io will be
done.