Added the memory dump function that will print all memory out into a clojure map. This map can now be read by the assembler to create a standalone vm cfile for the program.
The main routine now first looks to see if *main* is already defined.
The idea being that if the program is a proper memory dump that was
compiled to C that this function will be in memory already. If it is
this is what we want to run right away. While there are not any formal
tests to ensure all of this works, a simple file has been dumped and
compiled and runs as expected.

To create a memory dump right now it is best to write a program file of
all definitions and then end it with (dump). This will load all forms in
the file into memory and then print them. Usning > the output can be
sent to a file. When running the dump function and any other forms that
are not defines will not run, only those that are referenced through the
main function. So if a program is going to be compiled it should consist
only of defines. Though since they will not be run things like tests
could be put into the files an they could be run during the dump to
confirm program correctness before compiling it. Then once the program
was compiled and run they would be ignored, and not increase the size of
the dump since it outputs the entire memory. Also, once there is garbage
collection any memory used by these things would be collected. The only
downside is that the collecter would have to run sooner, because memory
was used that had no purpose. Though this doesn't matter until there is
actually a garbage collector.