Adjusted the dump to use only the number of spaces needed, rather than the whole memory. This uses the allocation pointer to the next empty memory. The issue that would come up is if there has been enough memory to cause garbage collection, then the pointer would be to an earlier place in memory. This is unlikely if a file with only defs is dumped, but cannot be totally ignored. The reason for not dumping all memory is so that file sizes will fit on kattis for some fun using the lisp to solve problems, but otherwise the file sized are still not very large compared to some.
Also, the dump process for a file is a bit cumbersome, because a file
with a main will run that after loading the forms. So once the dump is
done the program has to be run. This is a pain, because any program
output will end up in the file. So it may be useful to have another dump
that will halt and end.