Adjusted strings to be the previous char with a next pointer. This makes the string a much smaller memory object since the info and char are packed into a single word, and there is not pair like object pointing to it, but rather it just ponts to the next string head. This means there is really no such thing as a character but a single character string. Everything else is the same except the empty string now has 0 for the char byte and the null byte is not printed as anything. I have added some printing subroutines to print as is rather than in a readable way, but they will only be accessed by future print functions.
Made some string tests for operations with core functions like car and
cons, but since the string is no longer a pair like object these will
not work. It is difficult to decide if the same operations should be
made to work with strings, or new string operations to be created. At
the very least there needs to be some way to work with strings, so
probably the core list functions should be setup to work with strings.