Added new read functions to core that read the next form from the input stream and the next line.
The major issue right now is that read and read-num (not readln)
directly use the read subroutines and these read an extra character
before stopping. Since it might be that read is used to read several
forms from a line, if those forms were not separated by whitespace the
leading character could be lost. I am not sure what to do about this as
it is impossible to peek at the next char or put it back into the
stream. Most of the time these functions will be used to read well
structured data or just a number or symbol that should always be
separated by whitespace so it wont be an issue. But if it were used to
read data from a source that does not separate everything by whitespace
it could be a real problem.